% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRank.R
\name{convertRank}
\alias{convertRank}
\title{Convert rank values into pay grades}
\usage{
convertRank(data, ranks = "rankDat.RDS", payRanks = 4)
}
\arguments{
\item{data}{VADIR dataset}

\item{ranks}{Rank dataset, or path to rank dataset}

\item{payRanks}{Number of pay grades to use}
}
\value{
VADIR dataset with ranks converted to pay grade categories
}
\description{
Used in \code{sampleVADIR()} function. Probably not for user.
}
\examples{
\dontrun{
out <- convertRank(data, ranks = 'rankDat.RDS', payRanks = 4)
}
}
