% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Data import function to accommodate multiple filetypes}
\usage{
getData(filename, filetype = "csv", fixDates = FALSE, ...)
}
\arguments{
\item{filename}{Character string specifying the path to desired datafile}

\item{filetype}{Character string indicating filetype. Useful if no file
extension is provided in \code{filename}. If file extension is provided in
\code{filename} (recommended), then this argument is ignored. Accommodates
\code{"csv", "rds", "xlsx", "rdata", "sav", "txt"}}

\item{fixDates}{Logical. Determines whether to adjust date format.}

\item{...}{Additional arguments}
}
\value{
Imported datafile
}
\description{
Allows for easy data importation. Automatically detects filetype and applies
appropriate function for importing.
}
