% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStrata.R
\name{testStrata}
\alias{testStrata}
\title{Return correlations for demographics between population and sample}
\usage{
testStrata(out, data = NULL, metric = cor, zeros = FALSE)
}
\arguments{
\item{out}{Output of \code{\link{sampleVADIR}}}

\item{data}{Original VADIR data}

\item{metric}{Function for measuring similarity between population and sample}

\item{zeros}{Should empty strata be included?}
}
\value{
Similarity values for males and females
}
\description{
Used to evaluate the representativeness of the sample with regard to the
population. Males and females evaluated separately.
}
