% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VADIR_fake}
\alias{VADIR_fake}
\title{Fake VADIR data}
\format{
A data frame with ten variables, representing variables as they are
formatted within the actual VADIR database.
}
\usage{
VADIR_fake
}
\description{
Simulated VADIR data based solely on the variable names and appropriate
response options for each. Values of variables were generated based on
population proportions identified in a subsample of approximately 140,000
veterans from a version of the VADIR database obtained in 2020. However, this
simulated dataset does not fully represent population characteristics of
VADIR, and is simply meant as a faux tool for testing functions in the
\code{sampleVADIR} package.
}
\keyword{datasets}
